fix: unit test create and update data
